Rely on edX's runtime to show notifications
edX has a nice (though poorly documented!) APIs
for showing notifications to the user, which we've used
in the past and now return to in an expanded way.

Note that this commit is dependent upon a forthcoming
PR to the edX platform.

Other:

This also maps the escape key to runtime cancel.