automatically release to PyPI
This configures TravisCI to automatically build and upload `sdist` and
`bdist_wheel` packages to PyPI via our shared `appsembler` account.

Releases only go up when the change is pushed to the `master`
branch (TravisCI default; you don't want random PRs to be able to
trigger a release) with tags. The usual workflow for doing a release
to PyPI then is to update the version in `setup.py`, etc. Merge it to
`master`, tag a release locally (`git tag release-1.2.3`) then push
that up with `git push --tags`.